refactor: constant의 constants.js를 index.js로 파일명 변경
