+++ src/domain/LottoMachine.js
@@ -1,4 +1,4 @@
-import { LOTTO, GAME, PRIZE } from '../constant/constants.js';
+import { LOTTO, GAME, PRIZE } from '../constant/index.js';
 import randomGenerator from '../utils/randomGenerator.js';
 import Lotto from './Lotto.js';
