+++ src/validator/retry.js
@@ -1,4 +1,4 @@
-import { GAME, ERROR } from '../constant/constants.js';
+import { GAME, ERROR } from '../constant/index.js';
 
 export default function checkRetryFormat(command) {
   if (command !== GAME.RETRY && command !== GAME.EXIT) {
