+++ src/view/components/Amount.js
@@ -1,7 +1,7 @@
 import Component from './Component.js';
 import generationMessages from '../../utils/generationMessages.js';
 import Inputs from '../../utils/Inputs.js';
-import { LOTTO } from '../../constant/constants.js';
+import { LOTTO } from '../../constant/index.js';
 import LottoMachine from '../../domain/LottoMachine.js';
 
 export default class Amount extends Component {
